fix room

+++ database/migrations/2023_09_30_095133_create_booked_packages_detail_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('booked_packages_detail', function (Blueprint $table) {
             $table->id('id');
-            $table->integer('room_id');
-            $table->integer('booking_id');
+            $table->bigInteger('room_id');
+            $table->bigInteger('booking_id');
             $table->date('check_in_date');
             $table->date('check_oute_date');
             $table->string('status');
